backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
httpx
pydantic
openai
passlib
python-jose[cryptography]
email-validator
